requirements.txt
# requirements.txt
# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# AI/ML for summarization
openai==1.58.1
groq

# Web framework
fastapi==0.104.1
uvicorn==0.24.0

# Scheduling
APScheduler==3.10.4

# Frontend
streamlit==1.28.1
plotly==5.17.0

# Data processing
pandas==2.1.3
numpy==1.25.2

# Environment management
python-dotenv==1.0.0

# Logging and utilities
requests==2.31.0